+++ src/main/java/step2/LottoGame.java
@@ -1,13 +1,13 @@
 package step2;
 
-import step2.domain.Cache;
-import step2.domain.LottoMachine;
-import step2.domain.LottoNumber;
-import step2.domain.LottoStatistics;
-import step2.domain.RandomStrategy;
+import step2.domain.money.Cache;
+import step2.domain.lotto.LottoMachine;
+import step2.domain.lotto.LottoNumber;
+import step2.domain.lotto.LottoStatistics;
+import step2.domain.lotto.RandomStrategy;
 import step2.domain.ResultOfLottos;
-import step2.domain.User;
-import step2.domain.Wallet;
+import step2.domain.user.User;
+import step2.domain.user.Wallet;
 import step2.view.InputView;
 import step2.view.ResultView;
 import java.util.List;
