+++ src/main/java/step2/domain/ResultOfLottos.java
@@ -1,5 +1,6 @@
 package step2.domain;
 
+import step2.domain.lotto.LottoRank;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
