+++ src/main/java/step2/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2.domain.lotto;
 
 public class Lotto {
