+++ src/main/java/step2/domain/lotto/LottoStatistics.java
@@ -1,5 +1,8 @@
-package step2.domain;
+package step2.domain.lotto;
 
+import step2.domain.ResultOfLottos;
+import step2.domain.lotto.Lotto;
+import step2.domain.lotto.LottoNumber;
 import step2.util.CalcUtils;
 import java.util.List;
