+++ src/main/java/step2/domain/money/Money.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2.domain.money;
 
 public interface Money {
