+++ src/main/java/step2/domain/user/User.java
@@ -1,5 +1,9 @@
-package step2.domain;
+package step2.domain.user;
 
+import step2.domain.lotto.Lotto;
+import step2.domain.lotto.LottoMachine;
+import step2.domain.money.Cache;
+import step2.domain.money.Money;
 import java.util.List;
 
 public class User {
