+++ src/test/java/step2/domain/LottoTest.java
@@ -1,6 +1,8 @@
 package step2.domain;
 
 import org.junit.jupiter.api.Test;
+import step2.domain.lotto.Lotto;
+import step2.domain.lotto.LottoNumber;
 import java.util.Arrays;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
