+++ src/test/java/step2/domain/WalletTest.java
@@ -1,6 +1,9 @@
 package step2.domain;
 
 import org.junit.jupiter.api.Test;
+import step2.domain.money.Cache;
+import step2.domain.money.Money;
+import step2.domain.user.Wallet;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
